move utilities to separate files
